context tracks finished. Logic tracks favorites in progress
