+++ src/App.tsx
@@ -1,5 +1,6 @@
 import { BrowserRouter } from "react-router-dom";
 import { ThemeProvider } from "styled-components";
+import { TrackProvider } from "./Context/TracksContext";
 import { Router } from "./Router";
 import { GlobalStyles } from "./styles/global";
 import { defaultTheme } from "./styles/Themes/defaultTheme";
@@ -9,7 +10,9 @@ function App() {
   return (
     <ThemeProvider theme={defaultTheme}>
       <BrowserRouter>
-        <Router />
+        <TrackProvider>
+          <Router />
+        </TrackProvider>
         <GlobalStyles />
       </BrowserRouter>
