+++ src/Pages/Favorite/index.tsx
@@ -1,10 +1,9 @@
-import CardTrack from "../../Components/CardTrack";
 import { FavoriteContainer } from "./styles";
 
 export default function Favorite() {
   return (
     <FavoriteContainer>
-      <CardTrack />
+
 
     </FavoriteContainer>
   )
